% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq-color.R
\name{seq_color}
\alias{seq_color}
\title{Interpolate (non-linearly) between two colours}
\usage{
seq_color(
  col1,
  col2,
  n = 100,
  type = "cubic",
  direction = "in-out",
  colorspace = "hcl"
)
}
\arguments{
\item{col1, col2}{the two colours to interpolate between}

\item{n}{steps}

\item{type}{Valid types
are 'sine', 'quad', 'cubic', 'quart', 'quint', 'exp', 'circle', 'back',
'elastic', 'linear'. Default: 'cubic'}

\item{direction}{One of 'in', 'out or 'in-out'. Default: 'in-out'}

\item{colorspace}{Color space in which to do the interpolation. Default: 'hcl'
Can be any colorspace understood by the \code{farver} package i.e.
"cmy", "cmyk", "hsl", "hsb", "hsv", "lab" (CIE L*ab), "hunterlab" 
(Hunter Lab), "oklab", "lch" (CIE Lch(ab) / polarLAB), "luv", "rgb" 
(sRGB), "xyz", "yxy" (CIE xyY), "hcl" (CIE Lch(uv) / polarLuv), 
or "oklch" (Polar form of oklab).
Note: Not all color spaces make sense for interpolation.}
}
\value{
character vector containing a colour sequence
}
\description{
Interpolate (non-linearly) between two colours
}
\examples{
seq_color('red', 'blue', type = 'cubic', direction = 'in-out', colorspace = 'hcl')
}
