% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easings.R
\name{seq_ease}
\alias{seq_ease}
\title{Easing between two values}
\usage{
seq_ease(x1 = 0, x2 = 1, n = 100, type = "cubic", direction = "in-out")
}
\arguments{
\item{x1, x2}{Numeric values}

\item{n}{steps}

\item{type}{Valid types
are 'sine', 'quad', 'cubic', 'quart', 'quint', 'exp', 'circle', 'back',
'elastic', 'linear'. Default: 'cubic'}

\item{direction}{One of 'in', 'out or 'in-out'. Default: 'in-out'}
}
\value{
Numeric vector of length \code{n}
}
\description{
Easing between two values
}
\examples{
x <- seq_ease(x1 = 0, x2 = 1, n = 20, type = 'cubic', direction = 'in-out')
x
plot(x)
}
