% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easings.R
\name{seq_ease}
\alias{seq_ease}
\title{Easing between two values}
\usage{
seq_ease(x1 = 0, x2 = 1, n = 100, type)
}
\arguments{
\item{x1, x2}{Numeric values}

\item{n}{steps}

\item{type}{a combination of a name plus 'in', 'out' or 'in_out'. Valid names
are 'sine', 'quad', 'cubic', 'quart', 'quint', 'exp', 'circle', 'back',
'elastic'.  E.g. 'cubic-in-out', 'exp-in'}
}
\value{
Numeric vector of length \code{n}
}
\description{
Easing between two values
}
